Split filesystem tests from codebase tests
